[Refactor] 리뷰 반영 - profile로 환경 분리
